Extensionリファクタ
